add helper methods for type
